refactor: move FailedFiles under Runner namespace
Reviewer::FailedFiles → Reviewer::Runner::FailedFiles since it only
parses runner output. Update reference in Batch. Delete old file.
