+++ lib/reviewer.rb
@@ -15,7 +15,6 @@ require_relative 'reviewer/arguments'
 require_relative 'reviewer/batch'
 require_relative 'reviewer/capabilities'
 require_relative 'reviewer/command'
-require_relative 'reviewer/failed_files'
 require_relative 'reviewer/keywords'
 require_relative 'reviewer/output'
 require_relative 'reviewer/prompt'
