+++ lib/reviewer/batch.rb
@@ -69,7 +69,7 @@ module Reviewer
     # internally when scanning for file paths. Only called for tools that failed
     # (exit status exceeded the tool's max_exit_status threshold).
     def store_failed_files(tool, runner)
-      files = FailedFiles.new(runner.stdout, runner.stderr).to_a
+      files = Runner::FailedFiles.new(runner.stdout, runner.stderr).to_a
       Reviewer.history.set(tool.key, :last_failed_files, files) if files.any?
     end
