+++ lib/reviewer/runner.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'runner/failed_files'
 require_relative 'runner/guidance'
 require_relative 'runner/result'
 require_relative 'runner/strategies/captured'
